finalize refactor
