+++ src/AutoRest.CSharp/Common/AutoRest/Plugins/GeneratedCodeWorkspace.cs
@@ -10,6 +10,7 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using AutoRest.CSharp.Common.AutoRest.Plugins;
 using Azure.Core;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
@@ -174,346 +175,10 @@ namespace AutoRest.CSharp.AutoRest.Plugins
         public static bool IsCustomDocument(Document document) => !IsGeneratedDocument(document) && !IsSharedDocument(document);
         public static bool IsSharedDocument(Document document) => document.Folders.Contains(SharedFolder);
         public static bool IsGeneratedDocument(Document document) => document.Folders.Contains(GeneratedFolder);
-        private static bool IsMgmtRootDocument(Document document) => IsGeneratedDocument(document) && Path.GetDirectoryName(document.Name) is "Extensions" or "";
-
-        private class PublicDefinitionVisitor : CSharpSyntaxRewriter
-        {
-            public ImmutableHashSet<SyntaxNode> ModelDeclarations { get; private set; }
-            private readonly ImmutableHashSet<string> _modelsToKeep;
-
-            public PublicDefinitionVisitor(ImmutableHashSet<string> modelsToKeep)
-            {
-                _modelsToKeep = modelsToKeep;
-                ModelDeclarations = ImmutableHashSet<SyntaxNode>.Empty;
-            }
-
-            public PublicDefinitionVisitor()
-            {
-                _modelsToKeep = ImmutableHashSet<string>.Empty;
-                ModelDeclarations = ImmutableHashSet<SyntaxNode>.Empty;
-            }
-
-            public override SyntaxNode? VisitClassDeclaration(ClassDeclarationSyntax node)
-            {
-                node = (ClassDeclarationSyntax)base.VisitClassDeclaration(node)!;
-                if (IsPublic(node.Modifiers) && !_modelsToKeep.Contains(node.Identifier.ToString()))
-                {
-                    ModelDeclarations = ModelDeclarations.Add(node);
-                }
-
-                return node;
-            }
-
-            public override SyntaxNode? VisitStructDeclaration(StructDeclarationSyntax node)
-            {
-                node = (StructDeclarationSyntax)base.VisitStructDeclaration(node)!;
-
-                if (IsPublic(node.Modifiers) && !_modelsToKeep.Contains(node.Identifier.ToString()))
-                {
-                    ModelDeclarations = ModelDeclarations.Add(node);
-                }
-                return node;
-            }
-
-            public override SyntaxNode? VisitEnumDeclaration(EnumDeclarationSyntax node)
-            {
-                node = (EnumDeclarationSyntax)base.VisitEnumDeclaration(node)!;
-                if (IsPublic(node.Modifiers) && !_modelsToKeep.Contains(node.Identifier.ToString()))
-                {
-                    ModelDeclarations = ModelDeclarations.Add(node);
-                }
-                return node;
-            }
-        }
-
-        private class MemberVisitor : CSharpSyntaxRewriter
-        {
-            private List<SyntaxNode> _members = new();
-
-            public IEnumerable<SyntaxNode> PublicMembers => _members;
-
-            private Func<SyntaxTokenList, bool> _predicate;
-
-            public MemberVisitor(Func<SyntaxTokenList, bool> accessibility)
-            {
-                _predicate = accessibility;
-            }
-
-            public override SyntaxNode? VisitClassDeclaration(ClassDeclarationSyntax node)
-            {
-                node = (ClassDeclarationSyntax)base.VisitClassDeclaration(node)!;
-                if (_predicate(node.Modifiers))
-                {
-                    _members.Add(node);
-                }
-                return node;
-            }
-
-            public override SyntaxNode? VisitPropertyDeclaration(PropertyDeclarationSyntax node)
-            {
-                node = (PropertyDeclarationSyntax)base.VisitPropertyDeclaration(node)!;
-                if (_predicate(node.Modifiers))
-                {
-                    _members.Add(node);
-                }
-                return node;
-            }
-
-            public override SyntaxNode? VisitMethodDeclaration(MethodDeclarationSyntax node)
-            {
-                node = (MethodDeclarationSyntax)base.VisitMethodDeclaration(node)!;
-                if (_predicate(node.Modifiers))
-                    _members.Add(node);
-                return node;
-            }
-
-            public override SyntaxNode? VisitConstructorDeclaration(ConstructorDeclarationSyntax node)
-            {
-                node = (ConstructorDeclarationSyntax)base.VisitConstructorDeclaration(node)!;
-                if (_predicate(node.Modifiers))
-                    _members.Add(node);
-                return node;
-            }
-        }
-
-        private static bool IsPublic(SyntaxTokenList tokenList)
-            => tokenList.Any(token => token.IsKind(SyntaxKind.PublicKeyword));
-
-        private static bool IsStatic(SyntaxTokenList tokenList)
-            => tokenList.Any(token => token.IsKind(SyntaxKind.StaticKeyword));
-
-        private async Task<ImmutableHashSet<SyntaxNode>> GetAllPublicDeclaredModels(ImmutableHashSet<string> modelsToKeep)
-        {
-            var classVisitor = new PublicDefinitionVisitor(modelsToKeep);
-
-            foreach (var document in _project.Documents)
-            {
-                if (IsGeneratedDocument(document))
-                {
-                    var root = await document.GetSyntaxRootAsync();
-                    classVisitor.Visit(root);
-                }
-            }
-
-            return classVisitor.ModelDeclarations;
-        }
-
-        private static HashSet<string> _referenceAttributes = new HashSet<string> { "ReferenceType", "PropertyReferenceType", "TypeReferenceType" };
-
-        private bool IsReferenceType(SyntaxNode? root)
-        {
-            if (root is null)
-                return false;
-
-            var childNodes = root.DescendantNodes();
-            var typeNode = childNodes.OfType<TypeDeclarationSyntax>().FirstOrDefault();
-            if (typeNode is null)
-            {
-                return false;
-            }
-
-            var attributeLists = GetAttributeLists(typeNode);
-            if (attributeLists is null || attributeLists.Value.Count == 0)
-                return false;
-
-            foreach (var attributeList in attributeLists.Value)
-            {
-                if (_referenceAttributes.Contains(attributeList.Attributes[0].Name.ToString()))
-                    return true;
-            }
-
-            return false;
-        }
-
-        private SyntaxList<AttributeListSyntax>? GetAttributeLists(SyntaxNode node)
-        {
-            if (node is StructDeclarationSyntax structDeclaration)
-                return structDeclaration.AttributeLists;
-
-            if (node is ClassDeclarationSyntax classDeclaration)
-                return classDeclaration.AttributeLists;
-
-            return null;
-        }
-
-        private async Task<ImmutableHashSet<SyntaxNode>> GetRootNodes()
-        {
-            var classVisitor = new PublicDefinitionVisitor();
-            foreach (var document in _project.Documents)
-            {
-                var root = await document.GetSyntaxRootAsync();
-                // we add a declaration as root node when
-                // 1. the file is under `Generated` or `Generated/Extensions` which is handled by `IsMgmtRootDocument`
-                // 2. the declaration has a ReferenceType or similar attribute on it which is handled by `IsReferenceType`
-                // 3. the file is custom code (not generated and not shared) which is handled by `IsCustomDocument`
-                if (IsMgmtRootDocument(document) || IsReferenceType(root) || IsCustomDocument(document))
-                {
-                    classVisitor.Visit(root);
-                }
-            }
-
-            return classVisitor.ModelDeclarations;
-        }
-
-        private async IAsyncEnumerable<SyntaxNode> TraverseAllPublicModelsAsync(Compilation compilation, ImmutableHashSet<SyntaxNode> rootNodes)
-        {
-            var queue = new Queue<SyntaxNode>(rootNodes);
-            // traverse all the models starting from the root nodes
-            var visited = new HashSet<SyntaxNode>();
-            while (queue.Count > 0)
-            {
-                var node = queue.Dequeue();
-                if (visited.Contains(node))
-                    continue;
-                visited.Add(node);
-                // add this node to the public list
-                yield return node;
-                // add every type referenced by this node to the queue
-                await foreach (var child in GetReferencedTypes(compilation, node))
-                {
-                    queue.Enqueue(child);
-                }
-            }
-        }
-
-        private async IAsyncEnumerable<SyntaxNode> GetReferencedTypes(Compilation compilation, SyntaxNode root)
-        {
-            var semanticModel = compilation.GetSemanticModel(root.SyntaxTree);
-            var publicMemberVisitor = new MemberVisitor(IsPublic);
-            publicMemberVisitor.Visit(root);
-            foreach (var member in publicMemberVisitor.PublicMembers)
-            {
-                var symbol = semanticModel.GetDeclaredSymbol(member);
-                if (symbol == null)
-                    continue;
-                var list = new HashSet<SyntaxNode>();
-                await ProcessSymbol(symbol, list);
-                foreach (var node in list)
-                {
-                    yield return node;
-                }
-            }
-        }
-
-        private async Task ProcessSymbol(ISymbol? symbol, HashSet<SyntaxNode> result)
-        {
-            if (symbol == null || symbol.DeclaredAccessibility != Accessibility.Public)
-                return;
-            switch (symbol)
-            {
-                case INamedTypeSymbol typeSymbol:
-                    foreach (var reference in typeSymbol.DeclaringSyntaxReferences)
-                    {
-                        var node = await reference.GetSyntaxAsync();
-                        // short cut the recursive execution if this has already been in the result
-                        if (result.Contains(node))
-                            return;
-                        result.Add(node);
-                        if (HasDiscriminator(node, out var identifierCandidates))
-                        {
-                            // first find all the derived types from this type
-                            foreach (var derivedTypeSymbol in await SymbolFinder.FindDerivedClassesAsync(typeSymbol, _project.Solution))
-                            {
-                                if (identifierCandidates.Contains(derivedTypeSymbol.Name))
-                                {
-                                    await ProcessSymbol(derivedTypeSymbol, result);
-                                }
-                            }
-                        }
-                    }
-                    await ProcessSymbol(typeSymbol.BaseType, result);
-                    foreach (var typeArgument in typeSymbol.TypeArguments)
-                    {
-                        await ProcessSymbol(typeArgument, result);
-                    }
-                    break;
-                case IMethodSymbol methodSymbol:
-                    await ProcessSymbol(methodSymbol.ReturnType, result);
-                    foreach (var parameter in methodSymbol.Parameters)
-                    {
-                        await ProcessSymbol(parameter.Type, result);
-                    }
-                    break;
-                case IPropertySymbol propertySymbol:
-                    await ProcessSymbol(propertySymbol.Type, result);
-                    break;
-                default:
-                    throw new InvalidOperationException($"Not implemented for symbol {symbol.GetType()}");
-            }
-        }
-
-        private static bool HasDiscriminator(SyntaxNode node, [MaybeNullWhen(false)] out ImmutableHashSet<string> identifiers)
-        {
-            identifiers = null;
-            // only class models will have discriminators
-            if (node is ClassDeclarationSyntax classDeclaration)
-            {
-                var methodName = $"Deserialize{classDeclaration.Identifier.Text}";
-                var deserializeMethod = classDeclaration.Members.OfType<MethodDeclarationSyntax>().FirstOrDefault(method => method.Identifier.Text.Equals(methodName));
-                if (deserializeMethod != null && deserializeMethod.Body != null)
-                {
-                    var switchStatementNode = deserializeMethod.Body.DescendantNodes().OfType<SwitchStatementSyntax>();
-                    foreach (var switchNode in switchStatementNode)
-                    {
-                        if (switchNode.Expression.ToFullString().Contains("discriminator.GetString"))
-                        {
-                            // get the identifiers in the switch statements, some of them are class names of the derived classes
-                            // but since we do not have the full semantic model here, we cannot directly get which of them is
-                            // therefore here we just return all of them, and see if the derived type name shows in this list
-                            var identifierNodes = switchNode.Sections.SelectMany(section => section.DescendantNodes().OfType<IdentifierNameSyntax>());
-                            identifiers = identifierNodes.Select(identifier => identifier.Identifier.ToFullString()).ToImmutableHashSet();
-                            return true;
-                        }
-                    }
-                }
-                return false;
-            }
-
-            return false;
-        }
 
         public async Task InternalizeOrphanedModels(ImmutableHashSet<string> modelsToKeep)
         {
-            var compilation = await GetCompilationAsync();
-
-            // first get all the declared models
-            var definitions = await GetAllPublicDeclaredModels(modelsToKeep);
-            // get the root nodes
-            var rootNodes = await GetRootNodes();
-            // traverse all the root and recursively add all the things we met
-            var publicModels = TraverseAllPublicModelsAsync(compilation, rootNodes);
-            await foreach (var model in publicModels)
-            {
-                definitions = definitions.Remove(model);
-            }
-            // get the models we need to mark internal
-            foreach (var model in definitions)
-            {
-                MarkInternal(model);
-            }
-        }
-
-        private void MarkInternal(SyntaxNode declarationNode)
-        {
-            var newNode = declarationNode switch
-            {
-                MemberDeclarationSyntax declaration => MarkInternal(declaration),
-                _ => throw new InvalidOperationException()
-            };
-            var tree = declarationNode.SyntaxTree;
-            var document = _project.GetDocument(tree)!;
-            var newRoot = tree.GetRoot().ReplaceNode(declarationNode, newNode).WithAdditionalAnnotations(Simplifier.Annotation);
-            _project = _project.RemoveDocument(document.Id);
-            document = document.WithSyntaxRoot(newRoot);
-            _project = document.Project;
-        }
-
-        private static MemberDeclarationSyntax MarkInternal(MemberDeclarationSyntax memberDeclaration)
-        {
-            var publicTokenInList = memberDeclaration.Modifiers.First(token => token.IsKind(SyntaxKind.PublicKeyword));
-            var internalToken = SyntaxFactory.Token(publicTokenInList.LeadingTrivia, SyntaxKind.InternalKeyword, publicTokenInList.TrailingTrivia);
-            var newModifiers = memberDeclaration.Modifiers.Replace(publicTokenInList, internalToken);
-            return memberDeclaration.WithModifiers(newModifiers);
+            _project = await Internalizer.InternalizeAsync(_project, modelsToKeep);
         }
     }
 }
